use rclpy to get topic list
